# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: НейроКлип API — AI video production platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current user profile + balance + subscription
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CurrentUser" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      operationId: updateCurrentUser
      tags: [users]
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProfileRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CurrentUser" }
  /me/brand-kit:
    get:
      operationId: getBrandKit
      tags: [users]
      summary: Get my brand kit (defaults if not set)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BrandKit" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      operationId: updateBrandKit
      tags: [users]
      summary: Create or update brand kit (upsert)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BrandKit" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BrandKit" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /projects/{id}/scenes:
    post:
      operationId: addScene
      tags: [scenes]
      summary: Add a scene
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddSceneRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Scene" }
  /projects/{id}/scenes/{sceneId}:
    patch:
      operationId: updateScene
      tags: [scenes]
      summary: Update a scene
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: sceneId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSceneRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Scene" }
    delete:
      operationId: deleteScene
      tags: [scenes]
      summary: Delete a scene
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: sceneId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /projects/{id}/scenes/{sceneId}/regenerate-image:
    post:
      operationId: regenerateSceneImage
      tags: [scenes]
      summary: Regenerate image for a scene (mock)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: sceneId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Scene" }
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    UpdateProfileRequest:
      type: object
      properties:
        name: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
    CurrentUser:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        avatarUrl: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        role: { type: string, enum: [user, admin] }
        planId: { type: string }
        planName: { type: string }
        tokenBalance: { type: integer }
        videosUsedThisPeriod: { type: integer }
        videosQuota: { type: integer }
        currentPeriodEnd: { type: ["string", "null"], format: date-time }
        cancelAtPeriodEnd: { type: boolean }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - email
        - name
        - role
        - planId
        - planName
        - tokenBalance
        - videosUsedThisPeriod
        - videosQuota
        - cancelAtPeriodEnd
        - createdAt
    BrandKit:
      type: object
      properties:
        brandName: { type: string, maxLength: 60 }
        logoUrl: { type: ["string", "null"] }
        primaryColor: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
        accentColor: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
        fontChoice: { type: string, enum: [inter, manrope, rubik, playfair, montserrat] }
        watermarkText: { type: string, maxLength: 80 }
        tagline: { type: string, maxLength: 120 }
      required: [brandName, primaryColor, accentColor, fontChoice, watermarkText, tagline]
...
    AddSceneRequest:
      type: object
      properties:
        title: { type: string }
        narration: { type: string }
        imagePrompt: { type: string }
      required: [title, narration, imagePrompt]
    UpdateSceneRequest:
      type: object
      properties:
        title: { type: ["string", "null"] }
        narration: { type: ["string", "null"] }
        imagePrompt: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        animationType: { type: ["string", "null"] }
        transitionType: { type: ["string", "null"] }
        durationSec: { type: ["number", "null"] }
        orderIndex: { type: ["integer", "null"] }
    Scene:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        orderIndex: { type: integer }
        title: { type: string }
        narration: { type: string }
        imagePrompt: { type: string }
        imageUrl: { type: ["string", "null"] }
        audioUrl: { type: ["string", "null"] }
        durationSec: { type: number }
        animationType: { type: string }
        transitionType: { type: string }
      required: [id, projectId, orderIndex, title, narration, imagePrompt, durationSec, animationType, transitionType]
...